requirements.txt
opentelemetry-distro==0.52b1
opentelemetry-exporter-otlp==1.31.1
